requirements.txt
dnspython==2.6.1
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
